+++ src/main/java/com/serious/validation/support/RepeatMinValidator.java
@@ -31,7 +31,7 @@ public class RepeatMinValidator extends AbstractValidator {
     protected void doValidate(Object fieldValue, Object extensionValue, String errInfo) {
         Collection fieldValueCol = (Collection) fieldValue;
         String err = errInfo + "error with RepeatMin";
-        Preconditions.checkArgument(fieldValueCol.size() > (Integer) extensionValue, err);
+        Preconditions.checkArgument(fieldValueCol.size() > (Long) extensionValue, err);
     }
 
     @Override
